automate_scraping/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.41
botocore==1.38.41
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
duckduckgo_search==8.0.4
filelock==3.18.0
free_proxy==1.1.3
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-aws==0.2.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-mistralai==0.2.10
langchain-ollama==0.3.3
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.4.1
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
minify_html==0.16.4
mpire==2.10.2
multidict==6.5.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==1.44.0
nest-asyncio==1.6.0
numpy==2.3.1
ollama==0.5.1
openai==1.90.0
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
playwright==1.52.0
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
s3transfer==0.13.0
scrapegraph_py==1.12.0
scrapegraphai==1.58.0
semchunk==3.2.2
simpleeval==1.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.46.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
undetected-playwright==0.3.0
urllib3==2.5.0
yarl==1.20.1
zstandard==0.23.0
